% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analyse_CPUE_model-package.R
\docType{package}
\name{survival_analyse_CPUE_model-package}
\alias{survival_analyse_CPUE_model}
\alias{survival_analyse_CPUE_model-package}
\title{survival_analyse_CPUE_model: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/amelie-hst/survival_analyse_CPUE_model}
  \item Report bugs at \url{https://github.com/amelie-hst/survival_analyse_CPUE_model/issues}
}

}
\author{
\strong{Maintainer}: Amelie HOSTE \email{hoste@tourduvalat.org} (\href{https://orcid.org/0000-0002-4102-7985}{ORCID}) [copyright holder]

}
\keyword{internal}
